Add utilities for retrieving values from array
