+++ build.sh
@@ -1,9 +1,10 @@
 #! /bin/bash
 
-rm -rf ./build;
-mkdir -p ./build/lib;
+rm -rf ./bin;
+mkdir -p ./bin/lib;
 
 # Create object file
-clang -o ./build/lib/dynamicArray.o -c ./src/lib/dynamicArray.c;
+# clang -o ./bin/lib/dynamicArray.o -c ./src/lib/dynamicArray.c;
+clang -o ./bin/lib/dynamicArrayInt.o -c ./src/lib/dynamicArrayInt.c;
 
-clang -o ./build/utilities -Wall -Wextra -std=c2x -pedantic ./src/main.c ./build/lib/dynamicArray.o
+clang -o ./bin/utilities -Wall -Wextra -std=c2x -pedantic ./src/main.c ./bin/lib/dynamicArrayInt.o
