+++ run.sh
@@ -1,3 +1,3 @@
 #! /bin/bash
 
-./build/utilities
+./bin/utilities
