+++ src/lib/dynamicArray.h
@@ -15,12 +15,6 @@ typedef struct {
   size_t elementSize;
 } DynamicArray;
 
-typedef struct {
-  void *array;
-  size_t capacity;
-  size_t size;
-} DynamicArrayInt;
-
 // Free the array
 void freeArray(DynamicArray *a);
 
@@ -29,16 +23,5 @@ void insert(DynamicArray *a, void *element);
 
 // Initialise an Dynamic Array
 void initArray(DynamicArray *a, size_t initialSize, size_t elementSize);
-
-
-DynamicArrayInt initArrayInt(size_t initialCapacity);
-DynamicArrayInt initArrayInt();
-
-void growArrayInt(DynamicArrayInt *a);
-
-void insertIntIntoArray(DynamicArrayInt *a, int element); 
-
-void printDynamicArrayInt(DynamicArrayInt *a);
-
 
 #endif
