+++ src/main.c
@@ -1,12 +1,5 @@
 #include <stdio.h>
-#include "lib/dynamicArray.h"
-
-void printArray(DynamicArray *a) {
-  for (size_t i = 0; i < a->size; i++) {
-    printf("%lu ", *(int*)(char*)a->array + i * a->elementSize);
-  }
-  printf("\n");
-}
+#include "lib/dynamicArrayInt.h"
 
 int main() {
   printf("hello, world!!\n");
@@ -18,11 +11,8 @@ int main() {
   insertIntIntoArray(&a, 1);
   insertIntIntoArray(&a, 0);
   printDynamicArrayInt(&a);
-  // DynamicArray a;
-  // initArray(&a, 4, sizeof(int));
-  // printf("about to insert stuff\n");
-  // insert(&a, (void*)5);
-  // insert(&a, (void*)6);
-  // printArray(&a);
 
+  int ind = 0;
+  int wow = getIntFromArray(&a, ind);
+  printf("Value at index %d is %d", ind, wow);
 }
